requirements.txt
langchain
langchain-core
langchain-huggingface
transformers
huggingface-hub
python-dotenv
langchain-google-genai
google-generativeai
langchain-openai
openai
streamlit
langchain-community
pydantic
pydantic[email]